Should always use shell_out! unless trapping exit codes!
You are expected to trap the exit status and take action when using shell_out(),
without the !, other wise it is possible for the shell command to fail silently.
Failing silently is bad for many reasons, biggest of all it masks the real
issue.  The shell_out fails, but the recipe dies later.